+4я и 5я ДМ

+++ IPLabs/src/Pages/admin.php
@@ -48,7 +48,9 @@ if (isset($_SESSION['auth']) && isset($_SESSION['type']) && $_SESSION['type'] ==
                                                                                                 ?>" alt="avatar"></div>
                                     <div class="table-row-item"><?= $item['avatar_name'] ?></div>
                                     <a href="./edit-user.php?id=<?= $item['id'] ?>" class="manage-user" title="Edit"></a>
-                                    <a href="../server/delete-user.php?id=<?= $item['id'] ?>" class="delete-user" title="Delete"></a>
+                                    <?php if ($item['type'] != 0) { ?>
+                                        <a href="../server/delete-user.php?id=<?= $item['id'] ?>" class="delete-user" title="Delete"></a>
+                                    <?php } ?>
                                 </div>
                             <?php }
